# crates/larkstack/src/routes/mod.rs
//! The console HTTP surface: the gated admin API, the OAuth endpoints, the
//! generated OpenAPI spec + Scalar docs, and the SPA fallback.
//!
//! [`build`] wires every route into one [`axum::Router`]. Route registration
//! and OpenAPI collection are unified through [`utoipa_axum::router::OpenApiRouter`],
//! so the spec can never drift from the routes that actually exist.

use axum::{
    Json, Router,
    http::{StatusCode, header},
    response::{IntoResponse, Response},
    routing::get,
};
use serde::Serialize;
use utoipa::openapi::security::{ApiKey, ApiKeyValue, SecurityScheme};
use utoipa::{Modify, OpenApi, ToSchema};
use utoipa_axum::{router::OpenApiRouter, routes};
use utoipa_scalar::{Scalar, Servable};

use crate::HostState;

pub(crate) mod actions;
pub(crate) mod config;
pub(crate) mod console_auth;
pub(crate) mod events;
pub(crate) mod lark_apps;
pub(crate) mod oauth;
pub(crate) mod status;

/// Generic `{ "ok": true }` success body.
#[derive(Serialize, ToSchema)]
pub(crate) struct OkResponse {
    ok: bool,
}

impl OkResponse {
    pub(crate) fn ok() -> Self {
        Self { ok: true }
    }
}

/// Error body shared by every non-2xx JSON response.
#[derive(Serialize, ToSchema)]
pub(crate) struct ErrorResponse {
    error: String,
}

/// A typed API error: a status plus a message, rendered as [`ErrorResponse`].
pub(crate) struct ApiError {
    status: StatusCode,
    message: String,
}

impl ApiError {
    fn new(status: StatusCode, message: impl Into<String>) -> Self {
        Self {
            status,
            message: message.into(),
        }
    }
    pub(crate) fn bad_request(message: impl Into<String>) -> Self {
        Self::new(StatusCode::BAD_REQUEST, message)
    }
    pub(crate) fn not_found(message: impl Into<String>) -> Self {
        Self::new(StatusCode::NOT_FOUND, message)
    }
    pub(crate) fn unavailable(message: impl Into<String>) -> Self {
        Self::new(StatusCode::SERVICE_UNAVAILABLE, message)
    }
    pub(crate) fn internal(message: impl Into<String>) -> Self {
        Self::new(StatusCode::INTERNAL_SERVER_ERROR, message)
    }
}

impl IntoResponse for ApiError {
    fn into_response(self) -> Response {
        (
            self.status,
            Json(ErrorResponse {
                error: self.message,
            }),
        )
            .into_response()
    }
}

/// Parse the live config into an editable `toml_edit` document. Used by the
/// structured-edit handlers (`lark_apps`, `console_auth`, app enable) that
/// mutate one section while preserving the rest of the file — comments and all.
pub(crate) fn parse_doc(toml_str: &str) -> Result<toml_edit::DocumentMut, String> {
    toml_str
        .parse::<toml_edit::DocumentMut>()
        .map_err(|e| format!("config is not valid TOML: {e}"))
}

/// Persist an edited config document: write it to disk, then broadcast it on the
/// control plane so every affected supervisor restarts. The single write path
/// shared by all structured-edit handlers.
pub(crate) fn write_doc(s: &HostState, doc: toml_edit::DocumentMut) -> Result<(), String> {
    let body = doc.to_string();
    std::fs::write(s.config_path.as_path(), &body)
        .map_err(|e| format!("write {}: {e}", s.config_path.display()))?;
    s.control.set_config(std::sync::Arc::new(body));
    Ok(())
}

/// `GET /api/health` — ungated liveness probe.
#[utoipa::path(
    get, path = "/health", tag = "console",
    responses((status = 200, description = "Service is up", body = String)),
)]
pub(crate) async fn health() -> &'static str {
    "ok"
}

/// OpenAPI document root. Paths + component schemas are collected from the
/// [`OpenApiRouter`] in [`build`]; this only carries the metadata and the
/// session security scheme.
#[derive(OpenApi)]
#[openapi(
    info(
        title = "larkstack console API",
        description = "Admin API for the larkstack host: subsystem status, live config, \
                       an event stream (SSE), the Lark-app registry, action dispatch, and \
                       Lark OAuth login.",
        version = env!("CARGO_PKG_VERSION"),
    ),
    modifiers(&SessionScheme),
    tags(
        (name = "console", description = "Admin API — session-gated once OAuth is configured"),
        (name = "auth", description = "Lark OAuth login flow"),
    ),
)]
struct ApiDoc;

/// Declares the signed `lk_session` cookie as the `session` security scheme.
struct SessionScheme;

impl Modify for SessionScheme {
    fn modify(&self, openapi: &mut utoipa::openapi::OpenApi) {
        let components = openapi.components.get_or_insert_with(Default::default);
        components.add_security_scheme(
            "session",
            SecurityScheme::ApiKey(ApiKey::Cookie(ApiKeyValue::new("lk_session"))),
        );
    }
}

/// The OpenAPI-documented `/api` route set, minus `health` (which [`build`]
/// appends *after* the session gate). Shared with [`openapi_json`] so the
/// dumped spec can never drift from the routes that are served.
fn documented_api() -> OpenApiRouter<HostState> {
    OpenApiRouter::new()
        .routes(routes!(status::status))
        .routes(routes!(status::apps))
        .routes(routes!(config::get_config, config::put_config))
        .routes(routes!(config::set_app_enabled))
        .routes(routes!(config::set_app_lark_app))
        .routes(routes!(events::events))
        .routes(routes!(lark_apps::list, lark_apps::upsert))
        .routes(routes!(lark_apps::test))
        .routes(routes!(lark_apps::delete))
        .routes(routes!(
            console_auth::get_console_auth,
            console_auth::put_console_auth,
            console_auth::delete_console_auth
        ))
        .routes(routes!(actions::dispatch))
}

/// The ungated `/auth` OAuth route set.
fn documented_auth() -> OpenApiRouter<HostState> {
    OpenApiRouter::new()
        .routes(routes!(oauth::login))
        .routes(routes!(oauth::callback))
        .routes(routes!(oauth::logout))
        .routes(routes!(oauth::me))
}

/// The console's OpenAPI spec as pretty JSON, derivable without a running
/// host. `cargo xtask dump-openapi` writes it to `dashboard/openapi.json`,
/// which the frontend SDK is generated from.
pub fn openapi_json() -> String {
    let (_, spec) = OpenApiRouter::with_openapi(ApiDoc::openapi())
        .nest("/api", documented_api().routes(routes!(health)))
        .nest("/auth", documented_auth())
        .split_for_parts();
    spec.to_pretty_json().expect("OpenAPI spec serializes")
}

/// Assemble the full console router: gated `/api/*`, ungated `/auth/*`, the
/// OpenAPI spec + Scalar docs, and the embedded SPA fallback.
///
/// `app_routers` are per-App admin routers (each self-stated); they are mounted
/// at `/api/apps/<name>/` and inherit the same session gate. `ingress_routers`
/// are per-App public webhook routers, mounted at `/webhooks/<name>/` *outside*
/// the gate (callers authenticate with their own HMAC/token). Neither is part of
/// the OpenAPI spec.
pub(crate) fn build(
    state: HostState,
    app_routers: Vec<(String, Router)>,
    ingress_routers: Vec<(String, Router)>,
) -> Router {
    let gate = axum::middleware::from_fn_with_state(state.clone(), oauth::require_session);

    // `/api/*` — session-gated, except `/api/health` (added after `route_layer`).
    let mut api = documented_api();
    for (name, router) in app_routers {
        api = api.nest_service(&format!("/apps/{name}"), router);
    }
    let api = api.route_layer(gate).routes(routes!(health));

    // `/auth/*` — ungated.
    let auth = documented_auth();

    let (router, spec) = OpenApiRouter::with_openapi(ApiDoc::openapi())
        .nest("/api", api)
        .nest("/auth", auth)
        .split_for_parts();

    let spec_json = spec.to_pretty_json().expect("OpenAPI spec serializes");
    let docs: Router<HostState> = Scalar::with_url("/api/docs", spec).into();

    let mut router = router
        .route(
            "/api/openapi.json",
            get(move || {
                let spec = spec_json.clone();
                async move { ([(header::CONTENT_TYPE, "application/json")], spec) }
            }),
        )
        .merge(docs);

    // `/webhooks/<name>/*` — public inbound (webhooks, Lark event callbacks),
    // peer to `/api` and never under `route_layer(gate)`, so the session gate
    // doesn't touch them; each app authenticates its own callers.
    for (name, ingress) in ingress_routers {
        router = router.nest_service(&format!("/webhooks/{name}"), ingress);
    }

    router.fallback(crate::assets::serve).with_state(state)
}