# dashboard/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "larkstack console API",
    "description": "Admin API for the larkstack host: subsystem status, live config, an event stream (SSE), the Lark-app registry, action dispatch, and Lark OAuth login.",
    "contact": {
      "name": "S3anJia"
    },
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/config": {
      "get": {
        "tags": [
          "console"
        ],
        "summary": "`GET /api/config` — the current `config.toml`, verbatim.",
        "operationId": "get_config",
        "responses": {
          "200": {
            "description": "Current config.toml",
            "content": {
              "application/toml": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "put": {
        "tags": [
          "console"
        ],
        "summary": "`PUT /api/config` — validate, persist, and broadcast a replacement config.\nOnly the apps whose change key moved are restarted.",
        "operationId": "put_config",
        "requestBody": {
          "description": "Full replacement config.toml",
          "content": {
            "application/toml": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Saved; affected apps restart",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "description": "Body is not valid TOML",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Could not write the config file",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/config/{app}/enabled": {
      "put": {
        "tags": [
          "console"
        ],
        "summary": "`PUT /api/config/{app}/enabled` — flip whether the supervisor runs an app,\nwriting just `[<app>].enabled` and leaving the rest of the config (comments\nand all) untouched. A one-field affordance over the raw TOML editor for the\nmost common config edit; the affected app (re)starts or stops on broadcast.",
        "operationId": "set_app_enabled",
        "parameters": [
          {
            "name": "app",
            "in": "path",
            "description": "Registered app name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnabledReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Flag set; the app starts or stops",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnabledAck"
                }
              }
            }
          },
          "400": {
            "description": "That config section is not a table",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No such registered app",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Could not write the config file",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/config/{app}/lark-app": {
      "put": {
        "tags": [
          "console"
        ],
        "summary": "`PUT /api/config/{app}/lark-app` — bind (or clear) the `[lark-apps.<name>]`\nan app uses for Lark credentials, writing just `[<app>].lark_app` and leaving\nthe rest of the config untouched. A one-field affordance over the raw TOML\neditor; the affected app restarts on broadcast and resolves the new binding.\nA non-empty value must name a registered Lark app, else the app would fail\nits next build — so we reject it up front.",
        "operationId": "set_app_lark_app",
        "parameters": [
          {
            "name": "app",
            "in": "path",
            "description": "Registered app name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LarkAppReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Binding set (or cleared); the app restarts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LarkAppAck"
                }
              }
            }
          },
          "400": {
            "description": "lark_app is not a registered Lark app, or the section is not a table",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No such registered app",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Could not write the config file",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/console-auth": {
      "get": {
        "tags": [
          "console"
        ],
        "summary": "`GET /api/console-auth` — the current console sign-in binding.",
        "operationId": "get_console_auth",
        "responses": {
          "200": {
            "description": "Console sign-in binding",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsoleAuthView"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "put": {
        "tags": [
          "console"
        ],
        "summary": "`PUT /api/console-auth` — bind a Lark app as the console's OAuth client.\nSaving this enforces sign-in on the next request, so the bound app must\nalready carry working credentials in the registry (else the console would\nsilently stay open).",
        "operationId": "put_console_auth",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BindReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Binding saved; sign-in now required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "description": "lark_app is not a registered app with credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Could not write the config file",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "delete": {
        "tags": [
          "console"
        ],
        "summary": "`DELETE /api/console-auth` — unbind, reopening the console. The recovery path\nwhen an operator is about to lock themselves out (admins list omits them).",
        "operationId": "delete_console_auth",
        "responses": {
          "200": {
            "description": "Unbound; the console is open again",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "500": {
            "description": "Could not write the config file",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
    "/auth/callback": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "`GET /auth/callback` — exchange the code, check the allowlist, mint a\nsession.",
        "operationId": "callback",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "description": "Authorization code from Lark (present on success).",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "description": "Opaque state echoed back; must match the login handshake.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "error",
            "in": "query",
            "description": "Error code from Lark (present when the user denied access).",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "303": {
            "description": "Signed in; redirect to the console"
          },
          "400": {
            "description": "Missing/mismatched login state, or a Lark error"
          },
          "403": {
            "description": "Account is not in [console].admins"
          },
          "502": {
            "description": "Token exchange or user_info request failed"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BindReq": {
        "type": "object",
        "required": [
          "lark_app"
        ],
        "properties": {
          "admins": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Admin email allowlist; empty = any tenant user may sign in."
          },
          "lark_app": {
            "type": "string",
            "description": "Name of a registered `[lark-apps.<name>]` to sign users in with."
          },
          "redirect_uri": {
            "type": [
              "string",
              "null"
            ],
            "description": "Override the auto-derived `<host>/auth/callback` (must match what the\nLark app registers)."
          },
          "scope": {
            "type": [
              "string",
              "null"
            ],
            "description": "Extra OAuth scopes, space-separated (usually none)."
          }
        }
      },
      "ConsoleAuthView": {
        "type": "object",
        "description": "Body of `GET /api/console-auth`.",
        "required": [
          "configured",
          "admins"
        ],
        "properties": {
          "admins": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Admin email allowlist (empty = any tenant user may sign in)."
          },
          "configured": {
            "type": "boolean",
            "description": "Whether sign-in is currently enforced (a Lark app with working\ncredentials is bound). When false the console is OPEN."
          },
          "lark_app": {
            "type": [
              "string",
              "null"
            ],
            "description": "The bound `[lark-apps.<name>]`, if any."
          },
          "redirect_uri": {
            "type": [
              "string",
              "null"
            ]
          },
          "scope": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "EnabledAck": {
        "type": "object",
        "required": [
          "ok",
          "name",
          "enabled"
        ],
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "ok": {
            "type": "boolean"
          }
        }
      },
      "EnabledReq": {
        "type": "object",
        "required": [
          "enabled"
        ],
        "properties": {
          "enabled": {
            "type": "boolean"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "description": "Error body shared by every non-2xx JSON response.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "LarkAppAck": {
        "type": "object",
        "required": [
          "ok",
          "name"
        ],
        "properties": {
          "lark_app": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "ok": {
            "type": "boolean"
          }
        }
      },
      "LarkAppReq": {
        "type": "object",
        "properties": {
          "lark_app": {
            "type": [
              "string",
              "null"
            ],
            "description": "The `[lark-apps.<name>]` to bind for this app's Lark credentials, or\n`null`/empty to unbind (falling back to env / inline `[<app>.lark]`)."
          }
        }
      },
...
      "OkResponse": {
        "type": "object",
        "description": "Generic `{ \"ok\": true }` success body.",
        "required": [
          "ok"
        ],
        "properties": {
          "ok": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "lk_session"
      }
    }
  },
...
}